Bug Fix: client will now actually send survey data
Yep. It was that bad.

Introduced when I was fixing IE and Safari not having support for "for ... of".
And, it turns out, FormData is a special type of object that /allows/ you to
"for ... of" over it. So the i++ method of looping wasn't working as expected.
Awesome. Furthermore, even though Safari and IE "support" FormData, they don't
really support any of the useful features. Yay!

Ahem.

So, I'm just looping over the damn survey fields by hand and it seems to work.
Will test on IE and safari soon. But this should work. It's about as
lowest-common-denominator as you can get.

Close #31